move demos to examples directory
