+++ examples/demo_embedding.py
