+++ examples/demo_graph.py
